+++ examples/font.rs
@@ -21,7 +21,7 @@ use embedded_menu::{
 struct NavEvent;
 
 impl SelectValue for NavEvent {
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         // not part of the ASCII font
         "»"
     }
@@ -43,7 +43,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",