+++ examples/scrolling_single_touch.rs
@@ -22,7 +22,7 @@ enum NavEvent {
 }
 
 impl SelectValue for NavEvent {
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             Self::One => ">",
             Self::Two => "<-",
@@ -46,7 +46,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",