+++ examples/simple.rs
@@ -18,7 +18,7 @@ use embedded_menu::{
 struct NavEvent;
 
 impl SelectValue for NavEvent {
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         ">"
     }
 }
@@ -39,7 +39,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",